Filter Movies by Genre

This function filters the summer_movies dataset by the specified genre. If "All" is selected, it returns the entire dataset without filtering.

Usage

filter_movies_by_genre(data, genre)

Arguments

data

A data frame containing movie data, such as summer_movies.

genre

A character string specifying the genre to filter by. Use "All" to include all genres.

Value

A data frame filtered by the specified genre.

Examples

data("summer_movies")
# Filter for Comedy movies
filter_movies_by_genre(summer_movies, "Comedy")
#> # A tibble: 274 × 10
#>    tconst   title_type primary_title original_title  year runtime_minutes genres
#>    <chr>    <chr>      <chr>         <chr>          <dbl>           <dbl> <chr> 
#>  1 tt00267… movie      A Midsummer … A Midsummer N…  1935             133 Comed…
#>  2 tt00338… movie      The Teachers… Magistrarna p…  1941              86 Comedy
#>  3 tt00393… movie      One Swallow … En fluga gör …  1947              88 Comedy
#>  4 tt00415… movie      In the Good … In the Good O…  1949             102 Comed…
#>  5 tt00429… movie      Bountiful Su… Shchedroe leto  1951              87 Comed…
#>  6 tt00433… movie      Early Summer  Bakushû         1951             125 Comed…
#>  7 tt00472… movie      Summer's Clo… Nubes de vera…  1955              88 Comed…
#>  8 tt00486… movie      Smiles of a … Sommarnattens…  1955             109 Comed…
#>  9 tt00486… movie      Summertime    Summertime      1955             100 Comed…
#> 10 tt00494… movie      Love, Summer… Liebe, Sommer…  1956              96 Comed…
#> # ℹ 264 more rows
#> # ℹ 3 more variables: simple_title <chr>, average_rating <dbl>, num_votes <dbl>