Scatter plot function

This function plots a scatter graph given the data and the column names to visualize and also allows to edit the graphs axis labels and plot dimensions

plot_scatter_graph(
  data,
  x_axis_data,
  y_axis_data,
  x_axis_label,
  y_axis_label,
  title_label,
  text_size
)

Arguments

data

The data which should be visualized passed as a dataframe

x_axis_data

The column name of the passed data which should be displayed on the x axis passed as a dataframe which is the name of the column

y_axis_data

The column name of the passed data which should be displayed on the y axis passed as a dataframe which is the name of the column

x_axis_label

Label which will be displayed in the plot for the x axis passed as a string

y_axis_label

Label which will be displayed in the plot for the y axis passed as a string

title_label

Label which will be displayed on the top of the plot passed as a string

text_size

Text size applied for all text in the plot passed as a numeric value

Value

The generated plot with the applied parameters