Summer Movies Dataset

This dataset contains information about summer-themed movies, including movie titles, release year, runtime, genres, IMDb rating, and the number of votes. It is used to analyze trends and patterns in movies that are commonly released during the summer season.

Usage

summer_movies

Format

A data frame with the following columns:

tconst

Character. Unique identifier for each movie.

title_type

Character. Type of the title (e.g., "movie", "short").

primary_title

Character. Main title of the movie.

original_title

Character. Original title of the movie.

year

Integer. Year the movie was released.

runtime_minutes

Numeric. Runtime of the movie in minutes.

genres

Character. Genres associated with the movie (e.g., "Comedy", "Drama").

average_rating

Numeric. IMDb average rating of the movie.

num_votes

Integer. Number of votes received on IMDb.

Source

IMDb dataset, curated for TidyTuesday.

Examples

data(summer_movies)
head(summer_movies)
#> # A tibble: 6 × 10
#>   tconst    title_type primary_title original_title  year runtime_minutes genres
#>   <chr>     <chr>      <chr>         <chr>          <dbl>           <dbl> <chr> 
#> 1 tt0011462 movie      Midsummer Ma… Midsummer Mad…  1920              60 Drama 
#> 2 tt0026714 movie      A Midsummer … A Midsummer N…  1935             133 Comed…
#> 3 tt0033864 movie      The Teachers… Magistrarna p…  1941              86 Comedy
#> 4 tt0037325 movie      Summer Storm  Summer Storm    1944             106 Crime…
#> 5 tt0038406 movie      Centennial S… Centennial Su…  1946             102 Histo…
#> 6 tt0038738 tvMovie    A Midsummer … A Midsummer N…  1946             150 Drama…
#> # ℹ 3 more variables: simple_title <chr>, average_rating <dbl>, num_votes <dbl>